pyproject.toml
[project]
name = "habit-tracker"
version = "0.1.0"
description = ""
authors = [
    {name = "Kla1mzZz", email = "kla1mzz16@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi (>=0.115.12,<0.116.0)",
    "sqlalchemy[asyncio] (>=2.0.40,<3.0.0)",
    "alembic (>=1.15.2,<2.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "pydantic-settings (>=2.8.1,<3.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "python-jose (>=3.4.0,<4.0.0)",
    "pydantic[email] (>=2.11.2,<3.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "orjson (>=3.10.16,<4.0.0)",
    "torch (>=2.6.0,<3.0.0)",
    "ruff (>=0.11.4,<0.12.0)",
    "tqdm (>=4.67.1,<5.0.0)",
]

[tool.poetry]
packages = [{include = "habit_tracker", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 88

[tool.ruff.lint]
select = ["Q"]
ignore = ["Q001"]

[tool.ruff.format]
quote-style = "single"